ensemble-train/pyproject.toml
[project]
name = "model-train"
version = "0.1.0"
description = ""
authors = [
    {name = "Brandon Taylor",email = "btaylor9273@gmail.com"}
]
requires-python = ">=3.12,<3.15"
dependencies = [
    "numba (>=0.64.0,<0.65.0)",
    "numpy (<2.4.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "scipy (>=1.17.1,<2.0.0)",
    "librosa (>=0.11.0,<0.12.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "pandas (<2.4.0)",
    "fastai (>=2.8.7,<3.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "tsai (>=0.4.1,<0.5.0)",
    "torch (>=2.7.1,<2.8.0)",
    "torchaudio (>=2.7.1,<2.8.0)",
    "joblib (>=1.5.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false